dom/tests/mochitest/general/test_showModalDialog.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=862918
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 862918</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for window.showModalDialog. **/

  // The modal window needs to touch Cu, which means that it needs
  // SpecialPowers. But given the semantics of modal windows, we have to
  // do some funny stuff with postMessage to set this up.
  window.onmessage = function(evt) {
    is(evt.data, 'dosetup', "message from modal window is correct");
    var win = SpecialPowers.wrap(evt.source);
    win.wrappedJSObject.SpecialPowers = SpecialPowers;
    SpecialPowers.setWrapped(win.wrappedJSObject, 'is', SpecialPowers.wrap(is));
    SpecialPowers.setWrapped(win.wrappedJSObject, 'ok', SpecialPowers.wrap(ok));
    win.wrappedJSObject.go();
  };

  SimpleTest.waitForExplicitFinish();

  SpecialPowers.pushPrefEnv({ set: [[ "dom.disable_window_showModalDialog", false ]] }, doTest);
  async function doTest() {
    let iframe = document.createElement("iframe");
    document.body.appendChild(iframe);
    await new Promise(resolve => {
      iframe.addEventListener("load", resolve);
    });
    window.showModalDialog = iframe.contentWindow.showModalDialog;

    var someObj = { foo: 42, bar: "hi"};
    var xurl = location.toString()
                       .replace('mochi.test:8888', 'example.org')
                       .replace('test_showModal', 'file_showModal');
    if (xurl.indexOf('?') != -1)
      xurl = xurl.substring(0, xurl.indexOf('?'));
    is(window.showModalDialog('file_showModalDialog.html'), "rv: undefined");
    is(window.showModalDialog(xurl), undefined);
    is(window.showModalDialog('file_showModalDialog.html', 3), "rv: 3");
    is(window.showModalDialog(xurl, 3), undefined);
    is(window.showModalDialog('file_showModalDialog.html', someObj), "rv: " + someObj);
    is(window.showModalDialog(xurl, someObj), undefined);

    // Test sequential navigations.
    is(window.showModalDialog('file_showModalDialog.html?http://mochi.test:8888', 4),
                       'rv: 4');
    is(window.showModalDialog('file_showModalDialog.html?http://example.com', 4), undefined);
    is(window.showModalDialog('file_showModalDialog.html?http://example.com,http://mochi.test:8888', 4), 'rv: 4');

    // This test used to assert after gc. Make sure it doesn't.
    SpecialPowers.gc();
    SimpleTest.finish();
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=862918">Mozilla Bug 862918</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
